+++ vkconfig/CHANGELOG.md
@@ -26,6 +26,7 @@
 - Add link to [Vulkan Guide layers](https://github.com/KhronosGroup/Vulkan-Guide/blob/master/chapters/development_tools.md#vulkan-layers) list within the help menu
 - Update built-in VUIDs to the list included in SDK 1.2.176 #1511
 - Refactor the layer window to expose more layer documentation #1519
+- Add shader caching setting to validation built-in UI #1522
 
 ## [Vulkan Configurator 2.3.0 for Vulkan SDK 1.2.176.0](https://github.com/LunarG/VulkanTools/tree/sdk-1.2.176.0) - Mai 2021